resolved some comments
