+++ src/safeposix/vmmap.rs
@@ -1,3 +1,6 @@
+use crate::fdtables;
+
+use super::cage::{syscall_error, Errno};
 use super::vmmap_constants::*;
 use std::io;
 use nodit::NoditMap;
@@ -67,7 +70,7 @@ impl VmmapEntry {
     // this is effectively whatever mode the file was opened with
     // we need this because we shouldnt be able to change filed backed mappings 
     // to have protections exceeding that of the file
-    fn get_max_prot(&self, cage_id: u64, virtual_fd: u64 -> i32 {
+    fn get_max_prot(&self, cage_id: u64, virtual_fd: u64) -> i32 {
 
         let wrappedvfd = fdtables::translate_virtual_fd(cage_id, virtual_fd as u64);
         if wrappedvfd.is_err() {
@@ -76,12 +79,12 @@ impl VmmapEntry {
         let vfd = wrappedvfd.unwrap();
 
         // Declare statbuf by ourselves 
-        let mut libc_statbuf: stat = unsafe { std::mem::zeroed() };
+        let mut libc_statbuf: libc::stat = unsafe { std::mem::zeroed() };
         let libcret = unsafe {
             libc::fstat(vfd.underfd as i32, &mut libc_statbuf)
         };
 
-        libc_statbuf.mode as i32
+        libc_statbuf.st_mode as i32
     }
 }
